requirements.txt
streamlit>=1.28.0
python-docx>=1.1.0
Pillow>=10.0.0
